Rename *_GPU versions of integrator files to ease includes and makefiles
